EKF: Use GPS published accuracy to set observation noise
The published horizontal accuracy from the GPS receiver is used to set the observation noise, but is limited to be no less than EKF2_GPS_P_NOISE and more than EKF2_NOAID_NOISE.
This allows the receiver to adapt to varying GPS quality, but will not allow the observation noise to become so large that the attitude errors are no longer constrained.
This means that the horizontal position state variances can be related to absolute position accuracy under most conditions.